refactor(minimal-worker): split v2 schemas into worker/schemas.py, rename my_slugs
Follow-up to the previous review round: v2-only classes go in
worker/schemas.py so the v1 push path doesn't have to know about pull-mode
types. The v1 schemas and v1/v2 shared types (PipelineResultsResponse,
PipelineConfigResponse, SourceImageResponse, ...) stay in api/schemas.py.

Also rename the awkward `my_slugs` in loop.py. The list comes directly from
`pipeline_choices` (the dict of stub pipelines exported by api.api), so
iterate it directly instead of copying keys into a local variable.
