+++ processing_services/minimal/api/schemas.py
@@ -293,65 +293,7 @@ class ProcessingServiceInfoResponse(pydantic.BaseModel):
     # )
 
 
-# -----------------------------------------------------------------------------
-# v2 async worker schemas
-#
-# Mirror of the relevant classes in ami/ml/schemas.py. Kept in this package so
-# both the v1 FastAPI side (push / `/process`) and the v2 worker side
-# (pull / polling) share a single source of truth. When Antenna evolves the
-# canonical schemas, keep these in sync — field-for-field parity matters for
-# correct JSON round-trips.
-# -----------------------------------------------------------------------------
-
-
-class PipelineResultsError(pydantic.BaseModel):
-    """Error result when pipeline processing fails for a single task."""
-
-    error: str
-    image_id: str | None = None
-
-
-class PipelineProcessingTask(pydantic.BaseModel):
-    """A single image task reserved from the async job queue.
-
-    `reply_subject` is the NATS subject Antenna ACKs on when the result comes
-    back — the worker must round-trip it verbatim in the matching
-    PipelineTaskResult.
-    """
-
-    id: str
-    image_id: str
-    image_url: str
-    reply_subject: str | None = None
-
-
-class TasksResponse(pydantic.BaseModel):
-    """Response body of `POST /api/v2/jobs/{id}/tasks/`."""
-
-    tasks: list[PipelineProcessingTask] = []
-
-
-class PipelineTaskResult(pydantic.BaseModel):
-    """Result of processing a single PipelineProcessingTask."""
-
-    reply_subject: str
-    result: PipelineResultsResponse | PipelineResultsError
-
-
-class ProcessingServiceClientInfo(pydantic.BaseModel):
-    """Identity metadata sent by a processing service worker.
-
-    A ProcessingService in the DB may have multiple physical workers running
-    simultaneously; this lets the server distinguish them. Fields are
-    intentionally open — processing services can send any useful key/value
-    pairs (hostname, software version, pod name, etc).
-    """
-
-    model_config = pydantic.ConfigDict(extra="allow")
-
-
-class AsyncPipelineRegistrationRequest(pydantic.BaseModel):
-    """Body for `POST /api/v2/projects/{id}/pipelines/` from an async processing service."""
-
-    processing_service_name: str
-    pipelines: list[PipelineConfigResponse] = []
+# v2 pull-mode schemas (PipelineProcessingTask, PipelineTaskResult,
+# ProcessingServiceClientInfo, AsyncPipelineRegistrationRequest, ...) live in
+# `processing_services/minimal/worker/schemas.py` since only the worker path
+# uses them. The v1 schemas above are shared by both push and pull.
